+++ src/contexts/AppContext.js
@@ -3,20 +3,14 @@ import { auth, firebaseDB } from "../firebase";
 import { get, ref, onValue} from "firebase/database"
 // https://firebase.google.com/docs/auth/web/manage-users
 import { onAuthStateChanged } from 'firebase/auth';
-import firebase from 'firebase/compat/app';
-import { useNavigate } from 'react-router-dom';
+
 
 const AppContext = createContext();
 
 export function AppProvider({ children }) {
-    const navigate = useNavigate();
     const [user, setUser] = useState({});
     const [userDBRef, setUserDBRef] = useState({});
 
-    function handleLogout() {
-        firebase.auth().signOut();
-        navigate('/login');
-    }
 
     useEffect(() => {
         onAuthStateChanged(auth, (user) => {
@@ -26,7 +20,7 @@ export function AppProvider({ children }) {
             setUserDBRef((prevState) => (ref(firebaseDB, `${user?.uid}`)));
 
             // if user is anonymous, set the uid to 'anonymous' to avoid multiple anon uids (1 account only)
-            if (user.isAnonymous === true) {
+            if (user?.isAnonymous === true) {
                 user.uid = 'anonymous';
             }
         })
@@ -34,7 +28,7 @@ export function AppProvider({ children }) {
 
 
     return (
-        <AppContext.Provider value={{ auth, firebaseDB, user, userDBRef, handleLogout }}>
+        <AppContext.Provider value={{ auth, firebaseDB, user, userDBRef }}>
             {children}
         </AppContext.Provider>
     )
